double to int

+++ npbenchmark-main/ACS_VRPTW/src/vrptw/Route.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 public class Route
 {
     double load;//데係쨌쓺넓潼좆
-    double distance;//데係쨌쓺悧낀똑
+    int distance;//데係쨌쓺悧낀똑
     double time;
     ArrayList <Integer> route = new ArrayList<>();
 }
